requirements.txt
astroid==2.4.1
autopep8==1.5.2
bookkeeper @ git+https://github.com/AdrienCos/bookkeeper
flake8==3.8.2
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
numpy==1.18.4
paho-mqtt==1.5.0
pandas==1.0.3
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.17
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
wrapt==1.12.1
zipp==3.1.0
